lstm_stock_pytorch/requirements.txt
torch>=2.0.1
pandas>=2.0.3
numpy>=1.24.4
PyYAML>=6.0
scikit-learn>=1.3.0
pytest>=7.4.0
matplotlib>=3.7.0
